requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain-huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq